+++ docs/conf.py
@@ -48,6 +48,7 @@ templates_path = ['_templates']
 # You can specify multiple suffix as a list of string:
 # source_suffix = ['.rst', '.md']
 source_suffix = '.md'
+no_underscore_emphasis = True
 
 # The encoding of source files.
 #source_encoding = 'utf-8-sig'