+++ CHANGES.md
@@ -7,6 +7,7 @@ Version 0.1.2 (2016-06-01)
 --------------------------
 
 * Add reST's `::` marker support
+* Add options to disable emphasis by underscore (`_` or `__`)
 
 Version 0.1.1 (2016-05-30)
 --------------------------